# backend/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API онлайн-сервиса для IT-школы.",
        "title": "skadi API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Вход для существующего юзера по логину и паролю.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Вход для юзера.",
                "operationId": "auth-login",
                "parameters": [
                    {
                        "description": "authBody",
                        "name": "authBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.authBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.User"
                        }
                    },
                    "400": {
                        "description": "неверный логин или пароль"
                    }
                }
            }
        },
        "/auth/private/logout": {
            "post": {
                "security": [
                    {
                        "JWTRefresh": []
                    }
                ],
                "description": "Выход юзера (помещение refresh токена юзера в черный список).",
                "tags": [
                    "auth"
                ],
                "summary": "Выход юзера.",
                "operationId": "auth-private-logout",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "403": {
                        "description": "токен в черном списке"
                    }
                }
            }
        },
        "/auth/private/obtain": {
            "post": {
                "security": [
                    {
                        "JWTRefresh": []
                    }
                ],
                "description": "Получение нового access токена по данному refresh токену из cookie.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Получение access токена.",
                "operationId": "auth-private-obtain",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "403": {
                        "description": "токен в черном списке"
                    }
                }
            }
        },
        "/class": {
            "get": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Получение списка групп со всеми данными (ID и имена препода и учеников).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "class"
                ],
                "summary": "Получение списка групп.",
                "operationId": "class-list",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "page pagination param",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "example": 5,
                        "description": "per page pagination param",
                        "name": "per-page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "F26",
                        "description": "substring to filter classes by name (case-insensitive)",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.listClassOut"
                        }
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Создание новой группы со всеми данными и включение в неё переданных студентов.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "class"
                ],
                "summary": "Создание новой группы. [Только админ]",
                "operationId": "class-create",
                "parameters": [
                    {
                        "description": "classBody",
                        "name": "classBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.classBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Class"
                        }
                    },
                    "400": {
                        "description": "неверный ученик | неверный преподаватель | преподаватель не найден"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "409": {
                        "description": "группа с введенным названием уже существует"
                    }
                }
            }
        },
        "/class/short": {
            "get": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Получение списка групп (только ID и названия).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "class"
                ],
                "summary": "Получение списка групп (кратко).",
                "operationId": "class-list-short",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "page pagination param",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "example": 5,
                        "description": "per page pagination param",
                        "name": "per-page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "F26",
                        "description": "substring to filter classes by name (case-insensitive)",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Class"
                            }
                        }
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    }
                }
            }
        },
        "/class/{id}": {
            "get": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Получение всех данных о группе с инфой о преподе и учениках (ID и полные имена).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "class"
                ],
                "summary": "Получение группы по id.",
                "operationId": "class-read",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID группы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Class"
                        }
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "404": {
                        "description": "группа не найдена"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Удаление группы по её id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "class"
                ],
                "summary": "Удаление группы по id. [Только админ]",
                "operationId": "class-delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID юзера",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Частичное обновление группы (только переданные поля) по её id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "class"
                ],
                "summary": "Обновление группы по id. [Только админ]",
                "operationId": "class-update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID группы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "updateBody",
                        "name": "updateBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_class_controller_http_v1.updateBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Class"
                        }
                    },
                    "400": {
                        "description": "неверный ученик | неверный преподаватель | преподаватель не найден"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "404": {
                        "description": "группа не найдена"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.Class": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "description": "class id",
                    "type": "integer"
                },
                "name": {
                    "description": "class name",
                    "type": "string"
                },
                "schedule": {
                    "description": "class schedule",
                    "type": "string"
                },
                "students": {
                    "description": "student objects list",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Profile"
                    }
                },
                "teacher": {
                    "description": "teacher object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Profile"
                        }
                    ]
                }
            }
        },
...
        "entity.Contact": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "entity.Pagination": {
            "type": "object",
            "required": [
                "page",
                "pages",
                "per_page",
                "total"
            ],
            "properties": {
                "page": {
                    "description": "current page",
                    "type": "integer",
                    "example": 1
                },
                "pages": {
                    "description": "total pages amount",
                    "type": "integer",
                    "example": 7
                },
                "per_page": {
                    "description": "per page objects",
                    "type": "integer",
                    "default": 10,
                    "example": 5
                },
                "total": {
                    "description": "total objects amount",
                    "type": "integer",
                    "example": 32
                }
            }
        },
        "entity.Profile": {
            "type": "object",
            "required": [
                "fullname"
            ],
            "properties": {
                "address": {
                    "description": "user address",
                    "type": "string"
                },
                "contact": {
                    "$ref": "#/definitions/entity.Contact"
                },
                "extra": {
                    "description": "extra data (for admin only)",
                    "type": "string"
                },
                "fullname": {
                    "description": "user full name",
                    "type": "string"
                },
                "id": {
                    "description": "profile id (equals to user id)",
                    "type": "integer"
                },
                "parent_contact": {
                    "$ref": "#/definitions/entity.Contact"
                }
            }
        },
...
        "entity.User": {
            "type": "object",
            "required": [
                "id",
                "profile",
                "role",
                "username"
            ],
            "properties": {
                "class": {
                    "description": "student's class",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Class"
                        }
                    ]
                },
                "id": {
                    "description": "user id",
                    "type": "integer"
                },
                "profile": {
                    "description": "user profile",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Profile"
                        }
                    ]
                },
                "role": {
                    "description": "admin, teacher or student",
                    "type": "string"
                },
                "username": {
                    "description": "user username",
                    "type": "string"
                }
            }
        },
...
        "internal_app_class_controller_http_v1.updateBody": {
            "description": "updateBody represents a data to update class.",
            "type": "object",
            "properties": {
                "name": {
                    "description": "new class name",
                    "type": "string",
                    "maxLength": 50,
                    "example": "F26-2"
                },
                "schedule": {
                    "description": "new class schedule",
                    "type": "string",
                    "maxLength": 100,
                    "example": "пн 15:00-16:00"
                },
                "students": {
                    "description": "IDs of students (updated list) in the class",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "teacher_id": {
                    "description": "new teacher id",
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
        "v1.authBody": {
            "description": "authBody represents a data for user auth (log in).",
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "user password",
                    "type": "string",
                    "maxLength": 40,
                    "minLength": 8,
                    "example": "qwerty123"
                },
                "username": {
                    "description": "user username",
                    "type": "string",
                    "maxLength": 50,
                    "example": "user1"
                }
            }
        },
        "v1.classBody": {
            "description": "classBody represents a data with class.",
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "class name",
                    "type": "string",
                    "maxLength": 50,
                    "example": "B26-1"
                },
                "schedule": {
                    "description": "class schedule",
                    "type": "string",
                    "maxLength": 100,
                    "example": "сб 18:00-19:00"
                },
                "students": {
                    "description": "student ID list",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "teacher_id": {
                    "description": "teacher id",
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
        "v1.listClassOut": {
            "description": "listClassOut represents a classes list and pagination params.",
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "description": "classes list",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Class"
                    }
                },
                "pagination": {
                    "description": "pagination params",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Pagination"
                        }
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWTAccess": {
            "description": "Access JWT-token. Cookie will automatic add after auth is done.",
            "type": "apiKey",
            "name": "access",
            "in": "cookie"
        },
        "JWTRefresh": {
            "description": "Refresh JWT-token. Cookie will automatic add after auth is done.",
            "type": "apiKey",
            "name": "refresh",
            "in": "cookie"
        }
    }
}